==> TWSM/Homeworks/2_Text_Preprocessing.ipynb <==
# Exercise 2. Text Preprocessing

from google.colab import drive
drive.mount('/content/drive')

In this exercise, we will be using the 20-Newsgroups dataset. This version of the dataset contains about 11k newsgroups posts from 20 different topics.

We will learn how to apply the following steps:

1. Import and examine data
2. Remove initial text metadata with regex
3. Remove numbers, punctuation, tabs and convert to lower case with gensim
4. Stopwords and short words removal
5. Stemming and lematization

# Import packages
import pandas as pd
import re
from gensim.parsing.preprocessing import STOPWORDS, strip_tags, strip_numeric, strip_punctuation, strip_multiple_whitespaces, remove_stopwords, strip_short, stem_text
import pickle
import en_core_web_sm
import nltk
nltk.download('stopwords')

from nltk.corpus import stopwords

# 1. Import and examine data

# Import dataset
df = pd.read_json('https://raw.githubusercontent.com/selva86/datasets/master/newsgroups.json')
df.head()

# Examine dataset
print('Possible tagret values:')
print(df.target_names.unique())
print(' ')
print('Class distribution:')
print(df.target_names.value_counts())

*The classes are almost uniformly distributed.*

# The first entry in the content field
print(df.content[0])

df.shape

*The data contains 11314 rows.*

# 2. Remove initial text metadata

# Remove unnecessary headers
#(8768|9875|2353)
data = [re.sub('(?m)^(From:|Article-I.D.:|Organization:|Lines:|Nntp-Posting-Host:|Distribution:|Reply-To:|X-Newsreader:|Expires:|\s*-+).*\n', '', sent, flags=re.I) for sent in df.content]
data = [re.sub('(Subject:|Summary:|Keywords:)', '', sent, flags=re.I) for sent in data]

#print(df.iloc[10]['content'])
print(data[0])

# 3. Remove numbers, punctuation, tabs and convert to lower case with gensim

# Remove numbers
data=[strip_numeric(sent) for sent in data]
print(data[0])

# Remove punctuation
data=[strip_punctuation(sent) for sent in data]
print(data[0])

# Remove multiple spaces
data=[strip_multiple_whitespaces(sent) for sent in data]
print(data[0])

# Transform all data to lower-case
data=[sent.lower() for sent in data]
print(data[0])

# 4. Stopwords and short words removal 

Here we will compare the stopwords in gensim and nltk.

#### Gensim

all_stopwords = STOPWORDS
print(sorted(all_stopwords))

#### NLTK

nltk_stopwords = set(stopwords.words('english'))
print(sorted(nltk_stopwords))

data=[remove_stopwords(sent) for sent in data]
print(data[0])

data=[strip_short(sent) for sent in data]
print(data[0])

# 5. Stemming and Lematization

data_stem=[stem_text(sent) for sent in data]
#data_stem=[strip_short(sent) for sent in data_stem]
print(data_stem[0])

# Initialize spacy 'en' model
nlp = en_core_web_sm.load()
# Parse the sentence using the loaded 'en' model object `nlp`
data_lem_base=[nlp(sent) for sent in data]
# Extract the lemma for each token and join
data_lem=[" ".join([token.lemma_ for token in sent]) for sent in data_lem_base]
print(data_lem[0])

# 6. Corpus storage

IMPORTNANT: store your preprocessed corpus as you don’t want to do this over and over again.

pickle.dump(data_stem, open("/content/drive/MyDrive/TWSM_Data/Stemmed.pkl", "wb"))
pickle.dump(data_lem, open("/content/drive/MyDrive/TWSM_Data/Lemma.pkl", "wb"))
