==> TWSM/Homeworks/3_Text_Representation.ipynb <==
# Exercise 3. Text Representation Part 1

In this exercise we will derive the document representation of the preprocessed (stemmed) newsgroups dataset. We will apply sklearn as well as gensim to derive the Bag-of-words document representation. Those two packages are the standard packages for deriving the  Bag-of-words document representation.

We will calculate the following representations for each package:

* Absolute frequencies
* Relative frequencies
* TF-IDF frequences

Finally, we will derive N-grams for the dataset.

# Import packages
import pickle
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.preprocessing import Binarizer
import pandas as pd
from gensim.corpora import Dictionary
from gensim.models import TfidfModel
import math

from google.colab import drive
drive.mount('/content/drive')

# Import dataset
data_stem=pickle.load(open("/content/drive/MyDrive/TWSM_Data/Stemmed.pkl", "rb"))
print(data_stem[0])
print(' ')

len(data_stem)

# 1. Bag-of-words with sklearn

Apply the sklearn tranformations by removing all words that appear in less than min_df of the documents and more than max_df of the documents.

# Set transformers
vec_bin=Binarizer() # One-hot
vec_abs= CountVectorizer(max_df=0.95, min_df=0.05) #Absolute frequency
vec_rel = TfidfVectorizer(max_df=0.95, min_df=0.05, use_idf=False, norm='l1') # Relative frequency
vec_tf=TfidfVectorizer(max_df=0.95, min_df=0.05, smooth_idf=False) #Tf-IDF frequency

# Tranform stemmed data
corpus_sk_abs=vec_abs.fit_transform(data_stem)
corpus_sk_bin=vec_bin.fit_transform(corpus_sk_abs)
corpus_sk_rel=vec_rel.fit_transform(data_stem)
corpus_sk_tf=vec_tf.fit_transform(data_stem)

Show words in dictionary

print('The number of features is: ',len(vec_tf.get_feature_names()))
print('')
print(vec_tf.get_feature_names())
print(vec_abs.get_feature_names())
print(vec_rel.get_feature_names())

type(corpus_sk_rel)

corpus_sk_bin.toarray()[0]

# Put the frequencies of the first document in a data frame for the first text in the corpus
df_1=pd.DataFrame({'keys':vec_tf.get_feature_names(), 'bin': corpus_sk_bin.toarray()[0],'abs':corpus_sk_abs.toarray()[0],'rel': corpus_sk_rel.toarray()[0],'tf': corpus_sk_tf.toarray()[0]})

# Show only those with existing words
df_1=df_1[df_1.bin>0]
#print(df_1[df_1.rel>0])
print(df_1)
df_1.shape

# Number of times car appears in all documents
sum(corpus_sk_bin.toarray()[:,29])

# idf of car
import numpy as np
np.log(11314/697)+1

# model output
vec_tf.idf_[29]

# tfidf weights are normalised
sum(corpus_sk_tf.toarray()[0,:]**2)

nor=math.sqrt(sum((corpus_sk_abs.toarray()[0,:]*vec_tf.idf_)**2))

nor

corpus_sk_abs.toarray()[0,29]*vec_tf.idf_[29]/nor

print(df_1.sort_values(by=['tf'], ascending=False)[0:10])

print(df_1.sort_values(by=['abs'], ascending=False)[0:10])

# 2. Bag-of-words with gensim

As opposed to sklearn, gensim requires the data to be already split into tokens.

corpus_gen=[doc.split() for doc in data_stem]
corpus_gen[0]

Gensim assigns to each token (word) found in the data a unique id. In this way, it is possible to use those ids instead of the words for all operations.

# Create a gensim dictionary 
id2word=Dictionary(corpus_gen)

# Remove common and rare words
id2word.filter_extremes(no_below=566, no_above=0.95)
                        
# Show the tokens and their ids
print(id2word.token2id)

len(id2word.token2id.keys())

# Print all features
print(id2word.token2id.keys())

# Show how many times tokes appear in all documents
print(id2word.dfs) #car appears 697 times

# Absolute frequencies
corpus_gen_abs=[id2word.doc2bow(doc) for doc in corpus_gen]

corpus_gen_abs[0]

# Relative frequencies
corpus_gen_rel=[[(token[0],(token[1]/sum(n for _, n in doc))) for token in doc] for doc in corpus_gen_abs]

                  
# Binary frequencies
corpus_gen_bin=[[(token[0],1) for token in doc] for doc in corpus_gen_abs]

# Tfidf frequencies
tfidf=TfidfModel(dictionary=id2word, normalize=True)

# Get sklearn

#from gensim.models.tfidfmodel import df2idf
#from math import e
#def d2f(docfreq, totaldocs):
#    return df2idf(docfreq, totaldocs, log_base=e, add=1.0)
#tfidf=TfidfModel(dictionary=id2word, normalize=True, wglobal=d2f)

corpus_gen_tf=[tfidf[id2word.doc2bow(doc)] for doc in corpus_gen]

# idf term for car (id=1)
math.log2(11314/697)

tfidf.idfs[1]

len(corpus_gen_tf[0])

sum([corpus_gen_tf[0][i][1]**2 for i in range(len(corpus_gen_tf[0]))])

# normalisation
nor2=math.sqrt(sum([(corpus_gen_abs[0][i][1]*tfidf.idfs[i])**2 for i in range(len(corpus_gen_abs[0]))]))

nor2

(corpus_gen_abs[0][1][1]*tfidf.idfs[1])/nor2

print('binary')
print(corpus_gen_bin[0])
print(' ')
print('absolute')
print(corpus_gen_abs[0])
print(' ')
print('relative')
print(corpus_gen_rel[0])
print(' ')
print('tf-idf')
print(corpus_gen_tf[0])

len(corpus_gen_abs[0])

k=[list(id2word.token2id.keys())[i] for i in range(len(corpus_gen_bin[0]))]
tf=[corpus_gen_tf[0][i][1] for i in range(len(corpus_gen_bin[0]))]
rel=[corpus_gen_rel[0][i][1] for i in range(len(corpus_gen_bin[0]))]
ab=[corpus_gen_abs[0][i][1] for i in range(len(corpus_gen_abs[0]))]
bi=[corpus_gen_bin[0][i][1] for i in range(len(corpus_gen_bin[0]))]


# Put the frequencies in a data farme for the first text in the corpus
df_12=pd.DataFrame({'keys':k, 'bin': bi,'abs':ab,'rel':rel,'tf': tf})
print(df_12)
df_12.shape

df_merge=pd.merge(df_12, df_1, on='keys')
df_merge

# 3. Ngrams

# Two-gram absolute transformer (min=max=2 words)
bigram_vectorizer = CountVectorizer(ngram_range=(2, 2), max_df=0.95, min_df=0.05)

# Trasform data
corpus_sk_bi=bigram_vectorizer.fit_transform(data_stem)
print(bigram_vectorizer.get_feature_names())

You can see that there are much less words than for one-grams. The reason is that combination of words come less rarely over all documents than single words. However, if you remove the document frequency filtering, you will get much more words than for one-grams.

corpus_sk_bi_df=pd.DataFrame(corpus_sk_bi.toarray(), columns=bigram_vectorizer.get_feature_names())
corpus_sk_bi_df.head()

corpus_sk_bi_df.max()

The maximum frequency is 4 which is much less than for one-grams.
