==> TWSM/Homeworks/4_Text_Representation_2.ipynb <==
# Exercise 4. Text Representation Part 2



In this exercise we will apply the following models to the stemmed data from Exercise 2:

1.   Word2Vec
2.   Doc2vec
3.   BERT

At the end, we will derive a corpus with each of them which can be used in downstream tasks such as classification and clustering (see next exercises).


!pip install transformers

# Import packages
import pickle
import pandas as pd
from gensim.models import Word2Vec
from scipy.spatial.distance import cosine
from gensim.models.doc2vec import Doc2Vec, TaggedDocument
import tensorflow as tf
import torch
from transformers import BertTokenizer, BertModel
from keras.preprocessing.sequence import pad_sequences
from torch.utils.data import TensorDataset, DataLoader, RandomSampler, SequentialSampler

## 0. Load data

from google.colab import drive
drive.mount('/content/drive')

# Import dataset
data_lemma=pickle.load(open("/content/drive/MyDrive/TWSM_Data/Lemma.pkl", "rb"))
print(data_lemma[0])

## 1. Word2Vec


In this section we will train the word2vec model on the lemmatized data. 


# Prepare the dataset for the word2vec model
corpus_gen=[doc.split() for doc in data_lemma]

# Train the model for embeddings of size 100 considering words appearing in more than 566 documents, default window=5
model = Word2Vec(corpus_gen, size=100, min_count=566)
model.save('word2vec.model')

print([i for i in sorted(model.wv.vocab.keys())])

# Embedding for 'car'
vector = model.wv['car']
vector

# Most similar representations to 'car' based on cosine similarity
model.wv.most_similar('car')

# Embeddings' arithmetics
model.wv.most_similar(positive=['bike', 'machine'], topn=1)

In the following we will derive the corpus. Note that word2vec (as opposed to doc2vec) generates one embedding for each word in the document. These then need to be aggregated at a document level. The simplest way is to determine the average over all words, but you can also use other aggregators.

# Document representation for the text
corpus_w2v=[[model.wv[word] for word in doc if word in model.wv.vocab.keys()] for doc in corpus_gen]
positive=[i for i in range(len(corpus_gen)) if len(corpus_w2v[i])>0]

corpus_w2v2=[corpus_w2v[i] for i in positive]
data_lemma2=[data_lemma[i] for i in positive]

# Document average representation
corpus_w2v_avg_clean=[sum(words)/len(words) for words in corpus_w2v2]

# This corpus can be used later in clustering and classification tasks
print(corpus_w2v_avg_clean[10])

len(corpus_w2v)

len(corpus_w2v_avg_clean)

len(data_lemma2)

# Most simlar words to the document based on average representation
# This can be used to evaluate different aggregation methods and also provides interpretation of the document representation
print([token for (token,_) in model.wv.similar_by_vector(corpus_w2v_avg_clean[0])])

# cosine similarity to other documents
result=[(1 - cosine(corpus_w2v_avg_clean[0],corpus_w2v_avg_clean[i])) for i in range(1,len(corpus_w2v_avg_clean))]
most_similar=data_lemma2[result.index(max(result))+1]
print(data_lemma2[0])
print('')
print(most_similar)

len(result)

# Corpus as data frame that can be used in downstream tasks such as classification
corpus_w2v_avg_df=pd.DataFrame(corpus_w2v_avg_clean)
corpus_w2v_avg_df.head()

len(corpus_w2v_avg_df)

pickle.dump(corpus_w2v_avg_df, open("/content/drive/MyDrive/TWSM_Data/WordtoVecModel.pkl", "wb"))

## 2. Doc2Vec

# Run doc2vec on the tagged texts
documents = [TaggedDocument(doc, [i]) for i, doc in enumerate(corpus_gen)]
model2 = Doc2Vec(documents, vector_size=100, min_count=566)

# Embedding for the first document
vector = model2.infer_vector(corpus_gen[0])
vector


# cosine similarity to other documents
result=[(1 - cosine(vector,model2.infer_vector(corpus_gen[i]))) for i in range(1,len(corpus_gen))]
most_similar=data_lemma[result.index(max(result))+1]

print(most_similar)

len(corpus_gen)

# Final corpus for classification
corpus_d2v=pd.DataFrame([model2.infer_vector(doc) for doc in corpus_gen])

corpus_d2v.head()

pickle.dump(corpus_d2v, open("/content/drive/MyDrive/TWSM_Data/DoctoVecModel.pkl", "wb"))

## 3. BERT


Confirm that GPU is detected:

# Get the GPU device name.
device_name = tf.test.gpu_device_name()

# The device name should look like the following:
if device_name == '/device:GPU:0':
    print('Found GPU at: {}'.format(device_name))
else:
    raise SystemError('GPU device not found')

Assign the GPU device to torch:

# If there's a GPU available...
if torch.cuda.is_available():    

    # Tell PyTorch to use the GPU.    
    device = torch.device("cuda")

    print('There are %d GPU(s) available.' % torch.cuda.device_count())

    print('We will use the GPU:', torch.cuda.get_device_name(0))

# If not...
else:
    print('No GPU available, using the CPU instead.')
    device = torch.device("cpu")

In order to apply BERT, we need to derive three data objects for the text data:
1. Add [CLS] at the beginning and [SEP] at the end of each text. [SEP] is a legacy from teh model training. The result for [CLS] is then used later as document representation for classification tasks.
2. Tokenize the texts using BERT tokenizer
3. Pad or truncate the text to the maximum length (maximum 512)
4. Map the remaining tokens to BERT dictionary 





# 1. Add [CLS] at the beginning and [SEP] at the end of each text.
sentences = ["[CLS] " + query + " [SEP]" for query in data_lemma]
print(sentences[0])

# 2. Tokenize the texts using BERT tokenizer
tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')
tokenized_texts = [tokenizer.tokenize(sent) for sent in sentences]
print(tokenized_texts[0])

# Show token IDs based on BERT's training
print(tokenizer.convert_tokens_to_ids(tokenized_texts[0]))

In order to determine the maximum sequence length, we look at the list statistics.

leng=[]
for t in tokenized_texts:
  leng.append(len(t))
df=pd.DataFrame(leng)
df.describe()

df.quantile([.95, .99])

# 3. Pad the text to the maximum length, max 512

# Pad sequences that are less than MAX_LEN, if more, remove from the end
sentences_padded = pad_sequences(tokenized_texts,  dtype=object,maxlen=412,  value='[PAD]', truncating="post",padding="post", return_tensors = 'pt')
print(sentences_padded[0])

#4. Map the tokens to BERT dictionary 
# Convert the tokens to their index numbers in the BERT vocabulary
sentences_converted = [tokenizer.convert_tokens_to_ids(s) for s in sentences_padded]
print(sentences_converted[0])

# Create attention masks
masks = []

# Create a mask of 1s for each token followed by 0s for padding
for seq in sentences_converted:
  seq_mask = [float(i>0) for i in seq]
  masks.append(seq_mask)

# 5. Generate embeddings

#Convert all of our data into torch tensors, the required datatype for our model

inputs = torch.LongTensor(sentences_converted)
masks = torch.LongTensor(masks)

inputs.size()

masks.size()

# Apply Pretrained model to the sentences
model = BertModel.from_pretrained('bert-base-uncased')

model.to(device)

# Set the batch size.  
batch_size = 16  

# Create the DataLoader.
prediction_data = TensorDataset(inputs, masks)
prediction_sampler = SequentialSampler(prediction_data)
prediction_dataloader = DataLoader(prediction_data, sampler=prediction_sampler, batch_size=batch_size)

result=[]
i=0
for batch in prediction_dataloader:
  #print(i)
  # Add batch to GPU
  batch = tuple(t.to(device) for t in batch)


  # Unpack the inputs from our dataloader
  b_input_ids, b_input_mask = batch

  # Telling the model not to compute or store gradients, saving memory and 
  
  with torch.no_grad():
      # Forward pass, calculate embeddings
      outputs = model(b_input_ids)

  embeddings = outputs.pooler_output #CLS embeddings for the batch

  # Move em to CPU
  embeddings = embeddings.detach().cpu().numpy()
  
  # Store predictions and true labels
  result.append(embeddings)
  i=i+1


print('    DONE.')

#708 batches*16 texts with embedding size 768

final=[]
for b in result:
   for e in b:
      final.append(e)

# Final corpus
corpus_bert_df=pd.DataFrame(final)
corpus_bert_df.head()

pickle.dump(corpus_bert_df, open("/content/drive/MyDrive/TWSM_Data/BertModel.pkl", "wb"))
